feat: added facilities for testing reset attempts in POSIX layer

+++ src/hal/posix/system.cpp
@@ -12,7 +12,14 @@
 using namespace std;
 using namespace teller::hal::system;
 
+/** Time when the system was booted at */
 static struct timespec lastBootAt;
+
+/** Flag to prevent the delivery of the next reset signal in unit tests */
+static bool shouldPreventNextReset = false;
+
+/** Number of reset attempts that were prevented */
+static size_t numResetsPrevented = 0;
 
 namespace teller::hal::system {
 
@@ -20,10 +27,14 @@ void init()
 {
     int retval = clock_gettime(CLOCK_MONOTONIC, &lastBootAt);
     assert(retval == 0);
+
+    numResetsPrevented = 0;
+    shouldPreventNextReset = false;
 }
 
 void destroy()
 {
+    shouldPreventNextReset = false;
 }
 
 std::uint32_t getTimeSinceBootMsec(void)
@@ -46,8 +57,23 @@ void delayMsec(uint32_t delay)
 /* LCOV_EXCL_START */
 void requestReset()
 {
-    kill(getpid(), SIGUSR1);
+    if (shouldPreventNextReset) {
+        shouldPreventNextReset = false;
+        numResetsPrevented++;
+    } else {
+        kill(getpid(), SIGUSR1);
+    }
 }
 /* LCOV_EXCL_STOP */
+
+void preventNextReset()
+{
+    shouldPreventNextReset = true;
+}
+
+size_t countPreventedResetAttempts(void)
+{
+    return numResetsPrevented;
+}
 
 }
